+++ src/App.tsx
@@ -4,8 +4,8 @@ import "./App.css";
 import { useRecoilValue } from "recoil";
 import { categoriesAtom } from "./atoms";
 
-import Header from "./components/common/header/Header";
-import Footer from "./components/common/Footer";
+import Header from "./components/common/Header/Header";
+import Footer from "./components/common/Footer/Footer";
 import { createGlobalStyle } from "styled-components";
 import Main from "./pages/Main";
 import ProductList from "./pages/ProductList";
@@ -67,7 +67,7 @@ function App() {
       <GlobalStyle />
       <div className="App">
         <BrowserRouter>
-          <Header categories={categories} />
+          <Header />
           <Routes>
             {loading && <Route path="/" element={<SkeletonCard />} />}
             {!loading && <Route path="/" element={<Main data={data} />} />}
