+++ src/components/common/header/Logo.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+type LogoProps = {
+  children?: React.ReactNode;
+  path?: string;
+  title: string;
+};
+
+function Logo({ title, path }: LogoProps) {
+  return (
+    <div>
+      <Link to="/">{title}</Link>
+    </div>
+  );
+}
+
+export default Logo;
